Se le agrego una base de datos al programa

+++ src/Zoologico/Animal.java
@@ -36,7 +36,7 @@ public class Animal {
 		this.habitat = habitat;
 	}
 
-	public float getCantidad() {
+	public int getCantidad() {
 		return cantidad;
 	}
